Fix ML integration for prediction functionality

+++ backend/sgms-backend/src/main/java/com/sgms/sgms_backend/dto/RoomResponse.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @AllArgsConstructor
 public class RoomResponse {
 
-    private Long roomId;
+    private Integer roomId;
     private String roomNumber;
 }
